+++ src/SourceCode.Chasm/TreeNode.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 
 namespace SourceCode.Chasm
 {
-    public struct TreeNode : IEquatable<TreeNode>, IComparable<TreeNode>
+    public partial struct TreeNode : IEquatable<TreeNode>, IComparable<TreeNode>
     {
         #region Constants
 
@@ -12,8 +11,6 @@ namespace SourceCode.Chasm
         public static TreeNode EmptyBlob { get; } = new TreeNode(null, Sha1.Empty, NodeKind.Blob);
 
         public static TreeNode EmptyTree { get; } = new TreeNode(null, Sha1.Empty, NodeKind.Tree);
-
-        public static Comparer DefaultComparer { get; } = new Comparer();
 
         #endregion
 
@@ -81,63 +78,6 @@ namespace SourceCode.Chasm
         #region IComparable
 
         public int CompareTo(TreeNode other) => DefaultComparer.Compare(this, other);
-
-        #endregion
-
-        #region Comparer
-
-        public sealed class Comparer : IEqualityComparer<TreeNode>, IComparer<TreeNode>
-        {
-            internal Comparer()
-            { }
-
-            public int Compare(TreeNode x, TreeNode y)
-            {
-                // Nodes are always sorted by Name first (see TreeNodeList)
-                var cmp = StringComparer.Ordinal.Compare(x.Name, y.Name);
-                if (cmp != 0) return cmp;
-
-                // Then by Sha1 (in order to detect duplicate)
-                cmp = Sha1.DefaultComparer.Compare(x.Sha1, y.Sha1);
-                if (cmp != 0) return cmp;
-
-                cmp = ((int)x.Kind).CompareTo((int)y.Kind);
-                return cmp;
-            }
-
-            public bool Equals(TreeNode x, TreeNode y)
-            {
-                if (x.Kind != y.Kind) return false;
-                if (x.Sha1 != y.Sha1) return false;
-                if (!StringComparer.Ordinal.Equals(x.Name, y.Name)) return false;
-
-                return true;
-            }
-
-            public int GetHashCode(TreeNode obj)
-            {
-                var h = 11;
-
-                unchecked
-                {
-                    h = h * 7 + (obj.Name == null ? 0 : StringComparer.Ordinal.GetHashCode(obj.Name));
-                    h = h * 7 + (int)obj.Kind;
-                    h = h * 7 + obj.Sha1.GetHashCode();
-                }
-
-                return h;
-            }
-
-            // TODO: Is this the best place for this method
-            // Delegate dispatch faster than interface dispatch (https://github.com/dotnet/coreclr/pull/8504)
-            public static int NameComparison(TreeNode x, TreeNode y)
-                => string.CompareOrdinal(x.Name, y.Name);
-
-            // TODO: Is this the best place for this method
-            // Delegate dispatch faster than interface dispatch (https://github.com/dotnet/coreclr/pull/8504)
-            public static int Sha1Comparison(TreeNode x, TreeNode y)
-                => Sha1.DefaultComparer.Compare(x.Sha1, y.Sha1);
-        }
 
         #endregion
